Create quick and dirty program to test use of MapAndLoad()
The ImageHlp.dll MapAndLoad() function has been known to erroneously
fail on some older versions of the iPXE and wimboot binaries (see
e.g. https://github.com/ipxe/ipxe/commit/7b976dd30 and
https://github.com/ipxe/wimboot/commit/670c7e2).

Create a very quick and dirty program to simply call MapAndLoad() on
an arbitrary list of binaries, in order to allow this function to be
tested in isolation from the rest of the Secure Boot signing process.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>